requirements.txt
flask>=1.0, <3.0
groq
langchain
langchain_groq
langchain_anthropic
python-dotenv
Requests
scipy
pipenv
google-cloud-logging
gunicorn<21.0,>=19.2.0
pyyaml